Backend/requirements.txt
asgiref==3.9.2
certifi==2025.10.5
charset-normalizer==3.4.3
dj-database-url==3.0.1
Django==5.2.7
djangorestframework==3.16.1
idna==3.10
numpy==2.3.3
pandas==2.3.3
pgvector==0.4.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
